refactor + download zip file
